pythonflask-master/requirements.txt
async-timeout==4.0.2
attrs==22.2.0
boto3==1.26.48
botocore==1.29.48
bytecode==0.14.0
cattrs==22.2.0
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
dateutils==0.6.12
ddsketch==2.0.4
ddtrace==1.7.0
envier==0.4.0
exceptiongroup==1.1.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Excel==0.0.7
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.17.3
lml==0.1.0
MarkupSafe==2.1.1
packaging==23.0
Pillow==9.4.0
protobuf==4.21.12
# psycopg2
py-firebase-dynamic-links==1.0.1
pycrypt==0.7.2
pycryptodome==3.16.0
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-webio==0.1.4
PyJWT==2.6.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
razorpay==1.3.0
redis==4.4.2
redispy==3.0.0
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
SQLAlchemy==1.4.46
tenacity==8.1.0
texttable==1.6.7
typing_extensions==4.4.0
urllib3==1.26.14
Werkzeug==2.2.2
xmltodict==0.13.0
pymysql==1.0.0
Flask-Migrate==2.5.2